Improved installation script for OpenCV

+++ pulp/vitis-vision/tools/install-opencv3.sh
@@ -1,28 +1,46 @@
 #!/usr/bin/env bash
 
-my_home_dir=$(pwd)
+check_env_var()
+{
+    variable_name=$1
+    variable_value=$2
+
+    if [[ -z "$variable_value" ]]; then
+      echo "[ERROR] <$variable_name> is not defined yet. Please correct this!"
+    else
+      echo "<$variable_name> defined at '$variable_value'"
+    fi
+}
+
+# Check on the definition of a Vitis Vision workspace
+check_env_var workspace_vitis_vision $workspace_vitis_vision
+
+# Check on the definition of an install path
+check_env_var install_path $install_path
 
 # Create install directory
-mkdir -p $my_home_dir/install
-cd $my_home_dir/install
+mkdir -p $install_path
+cd $install_path
 
-# Remove old versions
-sudo apt purge -y libopencv* python-opencv
-sudo apt autoremove -y
+if [ $UNIMORE -eq 1 ]; then
+    # Remove old versions
+    sudo apt purge -y libopencv* python-opencv
+    sudo apt autoremove -y
 
-sudo apt update -y
-sudo apt upgrade -y
+    sudo apt update -y
+    sudo apt upgrade -y
 
-# Build pre-requisites
-sudo apt install -y build-essential cmake pkg-config git
-sudo apt install -y libjpeg-dev libtiff5-dev libpng-dev
-sudo apt install -y libavcodec-dev libavformat-dev libswscale-dev libxvidcore-dev libx264-dev libxine2-dev
-sudo apt install -y libv4l-dev v4l-utils
-sudo apt install -y libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev 
-sudo apt install -y libgtk2.0-dev
-sudo apt install -y mesa-utils libgl1-mesa-dri libgtkgl2.0-dev libgtkglext1-dev
-sudo apt install -y libatlas-base-dev gfortran libeigen3-dev
-sudo apt install -y python2.7-dev python3-dev python-numpy python3-numpy
+    # Build pre-requisites
+    sudo apt install -y build-essential cmake pkg-config git
+    sudo apt install -y libjpeg-dev libtiff5-dev libpng-dev
+    sudo apt install -y libavcodec-dev libavformat-dev libswscale-dev libxvidcore-dev libx264-dev libxine2-dev
+    sudo apt install -y libv4l-dev v4l-utils
+    sudo apt install -y libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev 
+    sudo apt install -y libgtk2.0-dev
+    sudo apt install -y mesa-utils libgl1-mesa-dri libgtkgl2.0-dev libgtkglext1-dev
+    sudo apt install -y libatlas-base-dev gfortran libeigen3-dev
+    sudo apt install -y python2.7-dev python3-dev python-numpy python3-numpy
+fi
 
 # Download OpenCV
 wget -O opencv.zip https://github.com/opencv/opencv/archive/3.3.1.zip
@@ -31,8 +49,8 @@ wget -O opencv_contrib.zip https://github.com/opencv/opencv_contrib/archive/3.3.
 unzip opencv_contrib.zip && mv opencv_contrib-3.3.1 opencv_contrib
 
 # Create OpenCV build directory
-mkdir -p $my_home_dir/install/opencv/build
-cd $my_home_dir/install/opencv/build
+mkdir -p $install_path/opencv/build
+cd $install_path/opencv/build
 
 # Build OpenCV
 cmake -D CMAKE_BUILD_TYPE=RELEASE \
@@ -66,4 +84,4 @@ sudo sh -c 'echo '/usr/local/lib' > /etc/ld.so.conf.d/opencv.conf'
 sudo ldconfig
 
 # Back home
-cd $my_home_dir
+cd $workspace_vitis_vision
